Add IServiceCollection extensions
Now to add MiniProfiler it's: services.AddMiniProfiler();
